refactor: rename tests.rs to basic_test.rs
